refactor: solve modal circular dependency issue
